--- a/.npm/bin/lefthook
+++ b/.npm/bin/lefthook
@@ -1,0 +1,13 @@
+#!/bin/sh
+
+dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
+
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    $dir/lefthook-linux $@
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    $dir/lefthook-mac $@
+elif [[ "$OSTYPE" == "cygwin" ]] || [[ "$OSTYPE" == "msys" ]] || [[ "$OSTYPE" == "win32" ]]; then
+    $dir/lefthook-win.exe $@
+else
+    Unsupported OS
+fi