Using lefthook through the NPM package causes performance issues, because NPX is a bit slow to start and execute each time lefthook is invoked. This can especially being seen when doing rebases with dozens of commits.

Here's a basic benchmark between native binary and NPM:

```shell
# Native:
$ time lefthook version
0.7.2
lefthook version  0.01s user 0.01s system 85% cpu 0.014 total

# NPM:
$ time npx lefthook version
0.7.2
npx lefthook version  0.09s user 0.04s system 97% cpu 0.125 total
```

If you rebase some commits, it will take ~9 times longer to finish rebasing with NPM vs native binary.

This PR makes the hooks search for the final binaries by themselves, rather than using the `npx` command (which is now only used as a fallback).

Here's a benchmark comparing native binary, NPM before this PR, and NPM after this PR:

```shell
# Native:
$ time git rebase -i HEAD~3
Successfully rebased and updated refs/heads/my-branch-name.
git rebase -i HEAD~3  0.28s user 0.25s system 24% cpu 2.138 total

# NPM before this PR:
$ time git rebase -i HEAD~3
Successfully rebased and updated refs/heads/my-branch-name.
git rebase -i HEAD~3  2.50s user 0.95s system 70% cpu 4.887 total

# NPM after this PR:
$ time git rebase -i HEAD~3
Successfully rebased and updated refs/heads/my-branch-name.
git rebase -i HEAD~3  0.27s user 0.26s system 34% cpu 1.559 total
```

You can see this PR makes NPM as fast as the native binary.