@Arkweid 2020-09-18
@nesk 
I've got a series of errors when testing file `bin/lefthook`
```
./lefthook: Bad substitution
./lefthook: [[: not found
./lefthook: [[: not found
./lefthook: [[: not found
./lefthook: [[: not found
./lefthook: [[: not found
./lefthook: Unsupported: not found
```
Could you look at?
It works when I switch `sh` to `bash` except `Unsupported OS`, there should be `echo "Unsupported OS"`