@nesk 2020-09-21
Sorry, I should have been more careful. Indeed, we should switch to `bash` instead of `sh` because it was written with Bash in mind, and `Unsupported OS` should outputted so we should use `echo "Unsupported OS"`.

Do you want a PR for this?