@Arkweid 2020-10-06
@nesk could we switch syntax to `sh` like in other scripts? I think in some systems bash is not supported(some docker environments/windows). Is it possible?