@nesk 2020-10-06
After a bit of searching, unfortunately `sh` is not a specific command line interpreter, [its whatever the system wants to provide us with](https://www.quora.com/What-is-the-difference-between-Shell-scripting-and-Bash-Shell-scripting).

I think we have two ways to manage this:

- We add a `bash` requirement.
- We try to improve the script, shouldn't be too hard but we will need to know how to reproduce the problem. Could you describe me the environment to reproduce [what you posted earlier](https://github.com/Arkweid/lefthook/pull/146#issuecomment-695076730)?

I'm really sorry I messed up your codebase… 😞