@Arkweid 2020-10-09
@nesk Here a docker file for Discourse. I test new builds for this one.
```
# TEST DISCOURSE LOCALY
# sudo docker build -t discourse-lefthook-test .
# sudo docker run -it discourse-lefthook-test bash

FROM ruby:2.6.6-slim-buster

# Common dependencies
RUN apt-get update -qq \
  && DEBIAN_FRONTEND=noninteractive apt-get install -yq --no-install-recommends \
    build-essential \
    gnupg2 \
    curl \
    git \
  && apt-get clean \
  && rm -rf /var/cache/apt/archives/* \
  && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/* \
  && truncate -s 0 /var/log/*log

# Add NodeJS to sources list
RUN curl -sL https://deb.nodesource.com/setup_12.x | bash -

# Add Yarn to the sources list
RUN curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add - \
  && echo 'deb http://dl.yarnpkg.com/debian/ stable main' > /etc/apt/sources.list.d/yarn.list

RUN apt-get update -qq && DEBIAN_FRONTEND=noninteractive apt-get -yq dist-upgrade && \
  DEBIAN_FRONTEND=noninteractive apt-get install -yq --no-install-recommends \
    libpq-dev \
    nodejs \
    yarn

RUN git clone https://github.com/discourse/discourse.git --depth=1

WORKDIR /discourse

RUN gem update --system && \
    gem install bundler

RUN bundle install
RUN yarn install

# COPY lefthook .
```

Now apply changes for npm package:
```
cd /discourse/node_modules/@arkweid/lefthook/bin
// rewrite indes.js and create lefthook file with script. Don't forget to make it executable
chmod +x lefthook 

// Now return to discourse/ dir 
cd -
yarn lefthook run lints

// And got this:
root@a86cc5b35f62:/discourse# yarn lefthook run lints
yarn run v1.22.4
$ /discourse/node_modules/.bin/lefthook run lints
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 3: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: Bad substitution
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 5: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: [[: not found
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 7: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: [[: not found
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 9: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: [[: not found
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 9: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: [[: not found
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 9: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: [[: not found
/discourse/node_modules/@arkweid/lefthook/bin/lefthook: 12: /discourse/node_modules/@arkweid/lefthook/bin/lefthook: Unsupported: not found
```