@nesk 2020-10-13
See #154 for a bugfix 🙂